Cleaned up documentation and formatting of many classes to more closely conform to modern Java standards, use of Javadoc, etc.
Wrote new documentation for EscPragmaLex and EscPragmaParser, as large
portions of both classes were undocumented.

Removed bogus imports that are no longer necessary to work around
Javac 1.0 bugs.

Rewrote section comments that used large amounts of space and
asterisks to denote method sets; we're just using normal EOL comments
now.

Moved the opening brace of a class to column 0 to help Emacs fontify
(much) more efficiently.  The opening brace of methods is still in the
trailer position.

Removed double-asterisk-prefixed comments in code emitted from astgen
to conform with the rest of the rewritten docs.

Fixed two of the three problem test cases for Javafe (test1007 and
test6).  Test1103 still fails; it is just a problem with the generated
'sources' file and relative paths.