Added action to fix-perms rule to make sure all test scripts are executable.  Added generate-toplevel-archive action to release rules. Added some docs on generating and applying patches.  Added patch creation rule for Javafe envelope (forgot it originally).
Added Mac OS X Simplify build to toplevel archive---we'll have to have
a separate simplify archive in the future as soon as we make any
changes to simplify that necessitate a version number change.

Added "zero-length-files" rule that archives all files of size 0
because diff/patch does not handle them properly (at least, as far as
I can tell so far).