Moved DEBUGGER_FLAGS and JAVA_REMOTE_DEBUG into Makefile.defs.
Removed typecheckEscjava rule since it was complex, fragile, and
now redundant given the availability of the selftypecheck rule.

Added a selftypecheck rule for Javafe to mirror that of Escjava.
Split typecheck testing into selftypecheck and typecheckJDK,
the latter of which is in the "alltests" rule, the former in
the "test" rule.

Made the typecheckJDK commands look for the appropriate jar files
in a directory "jars" under each source directory.  This is documented
in Makefile.local.sample as well as under the (now closed) bug
report.  (Closes bug#142)

Removed the use of the -source 1.4 switch from the JDK <1.4
typechecking tests.  (Closes bug#134)